Menu navigates to all pages
