+++ frontend/src/History.jsx
@@ -1,0 +1,17 @@
+import Navbar from "./Components/Navbar";
+
+function App(){
+// Create a Title component that'll render an <h1> tag with some styles
+
+// Use Title and Wrapper like any other React component – except they're styled!
+return(
+
+  <div> 
+    Transaction History
+   <Navbar/>
+
+  </div>
+);
+}
+
+export default App;
