+++ frontend/src/Home.jsx
@@ -1,6 +1,10 @@
 import React, { useEffect, useState } from "react";
 import styled, { createGlobalStyle } from "styled-components";
 import {FaHistory, FaUserAlt, FaChartLine} from "react-icons/fa";
+import "./Home.css";
+import { NavLink } from "react-router-dom";
+import Navbar from "./Components/Navbar";
+
 function App() {
   const [windowDimension, setWindowDimension] = useState(null);
 
@@ -19,125 +23,13 @@ function App() {
 
   const isMobile = windowDimension <= 640;
 
-  return (
-    <Styles.Wrapper>
-      <CSSReset />
+return (
+  <div className = "home">
+  All Stocks
+  <Navbar/>
 
-      {isMobile ? (
-        <MobileNavbar.Wrapper>
-          <MobileNavbar.Items>
-            <MobileNavbar.Item>
-              <MobileNavbar.Icon>
-                <FaChartLine size={18}/>
-              </MobileNavbar.Icon>
-              Stocks
-            </MobileNavbar.Item>
-            <MobileNavbar.Item>
-              <MobileNavbar.Icon>
-                <FaUserAlt size={16}/>
-              </MobileNavbar.Icon>
-              User
-            </MobileNavbar.Item>
-            <MobileNavbar.Item>
-              <MobileNavbar.Icon>
-                <FaHistory size={16}/>                
-              </MobileNavbar.Icon>
-              History
-            </MobileNavbar.Item>
-          </MobileNavbar.Items>
-        </MobileNavbar.Wrapper>
-      ) : (
-        <Navbar.Wrapper>
-          <Navbar.Logo>Logo</Navbar.Logo>
-          <Navbar.Items>
-            <Navbar.Item>Stocks</Navbar.Item>
-            <Navbar.Item>User</Navbar.Item>
-            <Navbar.Item>History</Navbar.Item>
-          </Navbar.Items>
-        </Navbar.Wrapper>
-      )}
-    </Styles.Wrapper>
+  </div>
   );
 }
-
-const Styles = {
-  Wrapper: styled.main`
-    display: flex;
-    background-color: #eeeeee;
-    height: 100vh;
-  `
-};
-
-const Navbar = {
-  Wrapper: styled.nav`
-    flex: 1;
-
-    align-self: flex-start;
-
-    padding: 1rem 3rem;
-
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    background-color: white;
-  `,
-  Logo: styled.h1`
-    border: 1px solid gray;
-    padding: 0.5rem 1rem;
-  `,
-  Items: styled.ul`
-    display: flex;
-    list-style: none;
-  `,
-  Item: styled.li`
-    padding: 0 1rem;
-    cursor: pointer;
-  `
-};
-
-const MobileNavbar = {
-  Wrapper: styled(Navbar.Wrapper)`
-    position: fixed;
-    width: 100vw;
-    bottom: 0;
-
-    justify-content: center;
-  `,
-  Items: styled(Navbar.Items)`
-    flex: 1;
-    padding: 0 2rem;
-
-    justify-content: space-around;
-  `,
-  Item: styled(Navbar.Item)`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    font-size: 1.2rem;
-  `,
-  Icon: styled.span``
-};
-
-const CSSReset = createGlobalStyle`
-  *,
-  *::before, 
-  *::after {
-    margin: 0; 
-    padding: 0;
-    box-sizing: inherit;
-  }
-
-  html {
-    font-size: 62.5%; /*1rem = 10px*/
-    box-sizing: border-box;      
-  }  
-
-  body {
-    font-size: 1.4rem;
-    font-family: sans-serif;
-  }
-`;
 
 export default App;
